Fix automatic clipboard paste vulnerability
This patch removes any | characters from the subtitle string which is
going to be copied to clipboard on Windows, because they cannot be
reliably escaped and are interpreted as pipe operator.  Before this
change, it was possible to put `|notepad.exe` at the end of any subtitle
line in an external subtitle and have Notepad open.  While it is highly
unlikely that this vulnerability would ever be used for a real attack,
it still shouldn't be there.

Additionally, `chcp 65001 >null` removes `Active code page: 65001` flood
from stdout.