Update filename_factory.lua
Added a function that makes a filename for the clip. 
Previously I kept using the make_snapshot_filename function even for video clips but this caused some bugs as both video clips and snapshots may have the same webp extension. In particular it led to problems when updating last note with ctrl+m