Add option to add placeholder text to empty cards
If you have vocab cards and your first field is the vocab or its
translation, you aren't able to create cards, only modify existing ones.

With this change you can add a placeholder text to the first field of a
card, e.g. "TODO", that you can later search for and edit.