Add empty card placeholder (#135)
* Add function spacing

* Add option to add placeholder text to empty cards

If you have vocab cards and your first field is the vocab or its
translation, you aren't able to create cards, only modify existing ones.

With this change you can add a placeholder text to the first field of a
card, e.g. "TODO", that you can later search for and edit.

* Remove config option and update README

* Don't overwrite first field it not empty

* Check first field every time

* fix

* print error

* use local variable

* fix

---------

Co-authored-by: Ren Tatsumoto <tatsu@autistici.org>